Update ext.py
fixes issue #8

We introduce an additional flag during the loop that parses the help text that determines whether to preserve line breaks across multiple lines of the help text. This flag is enabled by encountering '\b' and disabled when encountering a blank line. This is consistent with the expected behaviour of \b in click.